@@ -1,12 +1,24 @@
-import NextAuth from "next-auth";
+import "next-auth"
+import { JWT } from "next-auth/jwt"
 
 declare module "next-auth" {
   interface Session {
     user: {
-      name?: string | null;
-      email?: string | null;
-      image?: string | null;
-      isAdmin?: boolean;
-    };
+      id?: string
+      name?: string
+      email?: string
+      image?: string
+      isAdmin: boolean
+    }
+  }
+  
+  interface User {
+    isAdmin?: boolean
+  }
+}
+
+declare module "next-auth/jwt" {
+  interface JWT {
+    isAdmin?: boolean
   }
 } 
